ADS-6602193 implement basic flows
ADS-6602193 implement CreateStudioRoleCommandHandlerTest

ADS-6602193 implement whole flow for User

ADS-6602193 implement whole flow for Studio

ADS-6602193 implement whole flow for Brand

ADS-6602193 fix coverage

ADS-6602193 implement whole flow for Permission

ADS-6602193 implement whole flow for Role

ADS-6602193 cover cache triggers with UT

ADS-6602193 cover 'add_brand_studios' functions with UT

ADS-6602193 cover 'add_user_role' and 'delete_user_role' functions with UT

ADS-6602193 cover 'create_studio_role' function with UT

ADS-6602193 cover 'set_studio_roles' function with UT

ADS-6602193 cover 'create_brand_role' function with UT

ADS-6602193 cover 'set_brand_roles' function with UT

ADS-6602193cover 'set_role_permissions' function with UT

ADS-6602193 cover 'save role' function with UT

ADS-6602193 create db functions

ADS-6602193 fix IDs type

+++ administration-service/database/changelog/0.0.1/01_init_fdw_tables.sql
@@ -1,17 +1,17 @@
 CREATE FOREIGN TABLE IF NOT EXISTS fdw.user(
-    id INT
+    id BIGINT
 )
     SERVER user_fdw_db
     OPTIONS (schema_name 'users', table_name 'user');
 
 CREATE FOREIGN TABLE IF NOT EXISTS fdw.brand(
-    id INT
+    id BIGINT
 )
     SERVER brand_fdw_db
     OPTIONS (schema_name 'brands', table_name 'brand');
 
 CREATE FOREIGN TABLE IF NOT EXISTS fdw.studio(
-    id INT
+    id BIGINT
 )
     SERVER studio_fdw_db
     OPTIONS (schema_name 'studios', table_name 'studio');
